Move matchmaking-related files in new package and refactor
