+++ app/src/main/java/ch/epfl/sweng/SDP/matchmaking/MatchmakingInterface.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP;
+package ch.epfl.sweng.SDP.matchmaking;
 
 public interface MatchmakingInterface {
